splitting up shapepage.rs
