+++ rnote-ui/src/meson.build
@@ -48,7 +48,14 @@ rnote_ui_sources = files(
     'penssidebar/eraserpage.rs',
     'penssidebar/mod.rs',
     'penssidebar/selectorpage.rs',
-    'penssidebar/shaperpage.rs',
+    'penssidebar/shaperpage/mod.rs',
+    'penssidebar/shaperpage/border_width_widget.rs',
+    'penssidebar/shaperpage/constraints_widget.rs',
+    'penssidebar/shaperpage/fill_colorpicker_widget.rs',
+    'penssidebar/shaperpage/shapeconfiguration_widget.rs',
+    'penssidebar/shaperpage/shaperbuilder_widget.rs',
+    'penssidebar/shaperpage/shaperstyle_widget.rs',
+    'penssidebar/shaperpage/stroke_colorpicker_widget.rs',
     'penssidebar/toolspage.rs',
     'penssidebar/typewriterpage.rs',
     'settingspanel/mod.rs',
